refactor(sendspin): split source.go; drop dead NewFileSource stub
Separate the shared AudioSource contract (source.go) from the built-in
test-tone implementation (source_testtone.go), and delete the FileSource /
NewFileSource placeholder, which only ever returned a 'not implemented'
error. Update the package-doc server example to use NewTestTone instead of
the removed stub.

Phase 1 of the three-repo split.
